+++ android/src/main/kotlin/com/lucasjosino/on_audio_query/query/OnAudiosOnlyQuery.kt
@@ -5,12 +5,13 @@ import android.content.ContentResolver
 import android.content.Context
 import android.net.Uri
 import android.provider.MediaStore
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lucasjosino.on_audio_query.extras.loadArtwork
+import com.lucasjosino.on_audio_query.utils.loadArtwork
 import com.lucasjosino.on_audio_query.types.checkAudiosOnlyType
+import com.lucasjosino.on_audio_query.types.songProjection
 import com.lucasjosino.on_audio_query.types.sorttypes.checkSongSortType
+import com.lucasjosino.on_audio_query.utils.getExtraInfo
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import kotlinx.coroutines.Dispatchers
@@ -27,34 +28,6 @@ class OnAudiosOnlyQuery : ViewModel() {
     private lateinit var sortType: String
     @SuppressLint("StaticFieldLeak")
     private lateinit var context: Context
-
-    //Query projection
-    @SuppressLint("InlinedApi")
-    private val projection = arrayOf(
-            MediaStore.Audio.Media.DATA, //
-            MediaStore.Audio.Media.DISPLAY_NAME,
-            MediaStore.Audio.Media._ID,
-            MediaStore.Audio.Media.SIZE,
-            MediaStore.Audio.Media.ALBUM,
-            MediaStore.Audio.Media.ALBUM_ARTIST,
-            MediaStore.Audio.Media.ALBUM_ID,
-//            MediaStore.Audio.Media.ALBUM_KEY,
-            MediaStore.Audio.Media.ARTIST,
-            MediaStore.Audio.Media.ARTIST_ID,
-//            MediaStore.Audio.Media.ARTIST_KEY,
-            MediaStore.Audio.Media.BOOKMARK,
-            MediaStore.Audio.Media.COMPOSER,
-            MediaStore.Audio.Media.DATE_ADDED,
-            MediaStore.Audio.Media.DURATION,
-            MediaStore.Audio.Media.TITLE,
-            MediaStore.Audio.Media.TRACK,
-            MediaStore.Audio.Media.YEAR,
-            MediaStore.Audio.Media.IS_ALARM,
-            MediaStore.Audio.Media.IS_MUSIC, // 17
-            MediaStore.Audio.Media.IS_NOTIFICATION,
-            MediaStore.Audio.Media.IS_PODCAST,
-            MediaStore.Audio.Media.IS_RINGTONE
-    )
 
     //querySongs and queryAudios together
     fun queryAudiosOnly(context: Context, result: MethodChannel.Result, call: MethodCall) {
@@ -77,7 +50,7 @@ class OnAudiosOnlyQuery : ViewModel() {
 
     //Loading in Background
     private suspend fun loadAudiosOnly() : ArrayList<MutableMap<String, Any>> = withContext(Dispatchers.IO) {
-        val cursor = resolver.query(uri, projection, isOnly, null, sortType)
+        val cursor = resolver.query(uri, songProjection, isOnly, null, sortType)
         val audiosOnlyList: ArrayList<MutableMap<String, Any>> = ArrayList()
         while (cursor != null && cursor.moveToNext()) {
             val audiosData: MutableMap<String, Any> = HashMap()
@@ -91,13 +64,9 @@ class OnAudiosOnlyQuery : ViewModel() {
             val art = loadArtwork(context, audiosData["album"].toString())
             if (art.isNotEmpty()) audiosData["artwork"] = art
 
-            //Getting displayName without [Extension]. GitHub request - https://github.com/LucasPJS/on_audio_query/issues/5
-            val file = File(audiosData["_data"].toString())
-            audiosData["_display_name_wo_ext"] = file.nameWithoutExtension
-            //Adding only the extension
-            audiosData["file_extension"] = file.extension
-            //Adding parent file (All the path before file)
-            audiosData["file_parent"] = file.parent.orEmpty()
+            //Extra information from song
+            val extraInfo = getExtraInfo(audiosData["_data"].toString())
+            audiosData.putAll(extraInfo)
 
             audiosOnlyList.add(audiosData)
         }