+++ android/src/main/kotlin/com/lucasjosino/on_audio_query/query/OnWithFiltersQuery.kt
@@ -7,8 +7,9 @@ import android.net.Uri
 import android.provider.MediaStore
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lucasjosino.on_audio_query.extras.loadArtwork
+import com.lucasjosino.on_audio_query.utils.loadArtwork
 import com.lucasjosino.on_audio_query.types.*
+import com.lucasjosino.on_audio_query.utils.getExtraInfo
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import kotlinx.coroutines.Dispatchers
@@ -73,13 +74,9 @@ class OnWithFiltersQuery : ViewModel() {
                 val art = loadArtwork(context, withFiltersData["album"].toString())
                 if (art.isNotEmpty()) withFiltersData["artwork"] = art
 
-                //Getting displayName without [Extension]. GitHub request - https://github.com/LucasPJS/on_audio_query/issues/5
-                val file = File(withFiltersData["_data"].toString())
-                withFiltersData["_display_name_wo_ext"] = file.nameWithoutExtension
-                //Adding only the extension
-                withFiltersData["file_extension"] = file.extension
-                //Adding parent file (All the path before file)
-                withFiltersData["file_parent"] = file.parent.orEmpty()
+                //Extra information from song
+                val extraInfo = getExtraInfo(withFiltersData["_data"].toString())
+                withFiltersData.putAll(extraInfo)
 
                 withFiltersList.add(withFiltersData)
             } else withFiltersList.add(withFiltersData)