+++ example/lib/main.dart
@@ -11,8 +11,6 @@ Copyright: © 2021, Lucas Josino. All rights reserved.
 =============
 */
 
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:on_audio_query/on_audio_query.dart';
 
@@ -26,56 +24,53 @@ class Songs extends StatefulWidget {
 }
 
 class _SongsState extends State<Songs> {
-  List<SongModel> songList = [];
-  int? version;
+  late List<SongModel> songList;
+  late DeviceModel deviceModel;
 
   @override
   void initState() {
     super.initState();
-    getVersion();
+    getDevice();
   }
 
-  getVersion() async {
-    version = await OnAudioQuery().getDeviceSDK();
-  }
+  getDevice() async => deviceModel = await OnAudioQuery().queryDeviceInfo();
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: Text("Songs"),
+          title: Text("OnAudioQueryExample"),
           elevation: 2,
+          actions: [
+            IconButton(
+              onPressed: () {
+                setState(() {});
+              },
+              icon: Icon(Icons.add),
+            )
+          ],
         ),
         body: FutureBuilder(
-          future: OnAudioQuery().querySongs(SongSortType.DEFAULT,
-              OrderType.ASC_OR_SMALLER, UriType.EXTERNAL, true),
+          // DEFAULT: SongSortType.TITLE, OrderType.ASC_OR_SMALLER, UriType.EXTERNAL and false
+          future: OnAudioQuery().querySongs(null, null, null, true),
           builder: (context, AsyncSnapshot<List<SongModel>> item) {
             if (item.data != null) {
               songList = item.data!;
               return ListView.builder(
                 itemCount: songList.length,
                 itemBuilder: (context, index) {
+                  print(songList[index].artwork);
                   return ListTile(
                     title: Text(songList[index].title),
                     subtitle: Text(songList[index].artist),
                     trailing: Icon(Icons.arrow_forward_rounded),
-                    leading: version! >= 29
-                        ? QueryArtworkWidget(
-                            id: songList[index].id,
-                            type: ArtworkType.AUDIO,
-                          )
-                        : ClipRRect(
-                            borderRadius: BorderRadius.circular(50),
-                            child: Image(
-                              image: FileImage(
-                                File(songList[index].artwork!),
-                              ),
-                              height: 50,
-                              width: 50,
-                              fit: BoxFit.cover,
-                            ),
-                          ),
+                    leading: QueryArtworkWidget(
+                      id: songList[index].id,
+                      type: ArtworkType.AUDIO,
+                      artwork: songList[index].artwork,
+                      deviceInfo: deviceModel,
+                    ),
                   );
                 },
               );