+++ lib/details/types/songs_by_type.dart
@@ -1,0 +1,13 @@
+part of on_audio_query;
+
+/// Defines a specific type for query audios.
+enum SongsByType {
+  /// Query songs based in a specific [ID].
+  ID,
+
+  /// Query songs based in a specific [TITLE].
+  TITLE,
+
+  /// Query songs based in a specific [DATA].
+  DATA,
+}