+++ lib/on_audio_query.dart
@@ -15,6 +15,7 @@ library on_audio_query;
 
 //
 import 'dart:async';
+import 'dart:io';
 import 'dart:typed_data';
 
 //Dart/Flutter
@@ -30,6 +31,8 @@ part 'details/models/album_model.dart';
 part 'details/models/artist_model.dart';
 part 'details/models/playlist_model.dart';
 part 'details/models/genre_model.dart';
+//
+part 'details/models/device_model.dart';
 
 //Sort Types
 part 'details/types/sort_types/song_sort_type.dart';
@@ -44,6 +47,7 @@ part 'details/types/audios_from_type.dart';
 part 'details/types/audios_only_type.dart';
 part 'details/types/with_filters_type.dart';
 part 'details/types/uri_type.dart';
+part 'details/types/songs_by_type.dart';
 
 //Widget
 part 'widget/queryArtworkWidget.dart';